+++ python-demo/config/configs.ini
@@ -12,7 +12,7 @@ quote_asset=USDT
 
 [pure arbitrage]
 strategy_name=Pure Arbitrage BTCUSDT Injective exchange Binance Demo
-priv_key= input private key here
+priv_key=5d386fbdbf11f1141010f81a46b40f94887367562bd33b452bbaa6ce1cd1381e
 binance_api_key= input binance api key
 binance_api_secret= input binance api secret
 is_mainnet=true
